+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 
 ### Added
 
+- New values to `set-cookie` and `set-cookie-reload` scriptlets: `emptyArr`, `emptyObj` [#497].
 - Ability to set random response content in `prevent-fetch` scriptlet [#416].
 - Ability to choose CSS injection method in `inject-css-in-shadow-dom` scriptlet [#477].
 - TypeScript types for CoreLibs provided [`ContentScriptApi`](./README.md#scriptlets-api--content-script-api).
@@ -33,6 +34,7 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 [#416]: https://github.com/AdguardTeam/Scriptlets/issues/416
 [#477]: https://github.com/AdguardTeam/Scriptlets/issues/477
 [#496]: https://github.com/AdguardTeam/Scriptlets/issues/496
+[#497]: https://github.com/AdguardTeam/Scriptlets/issues/497
 
 ## [v2.1.7] - 2025-04-03